+++ core/src/main/java/bisq/core/offer/CreateOfferService.java
@@ -182,7 +182,7 @@ public class CreateOfferService {
                 currencyCode,
                 makerFeeAsCoin);
 
-        OfferPayload offerPayload = new OfferPayload(offerId,
+        FeeTxOfferPayload offerPayload = new FeeTxOfferPayload(offerId,
                 creationTime,
                 makerAddress,
                 pubKeyRing,
