+++ core/src/main/java/bisq/core/trade/statistics/TradeStatistics2.java
@@ -21,6 +21,7 @@ import bisq.core.monetary.Altcoin;
 import bisq.core.monetary.AltcoinExchangeRate;
 import bisq.core.monetary.Price;
 import bisq.core.monetary.Volume;
+import bisq.core.offer.FeeTxOfferPayload;
 import bisq.core.offer.Offer;
 import bisq.core.offer.OfferPayload;
 import bisq.core.trade.Trade;
@@ -60,6 +61,7 @@ import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nullable;
 
+import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
@@ -93,7 +95,11 @@ public final class TradeStatistics2 implements ProcessOncePersistableNetworkPayl
         Offer offer = trade.getOffer();
         checkNotNull(offer, "offer must not ne null");
         checkNotNull(trade.getTradeAmount(), "trade.getTradeAmount() must not ne null");
-        return new TradeStatistics2(offer.getOfferPayload(),
+
+        checkArgument(offer.getOfferPayload() instanceof FeeTxOfferPayload);
+        FeeTxOfferPayload offerPayload = (FeeTxOfferPayload) offer.getOfferPayload();
+
+        return new TradeStatistics2(offerPayload,
                 trade.getTradePrice(),
                 trade.getTradeAmount(),
                 trade.getDate(),
@@ -137,7 +143,7 @@ public final class TradeStatistics2 implements ProcessOncePersistableNetworkPayl
     @JsonExclude
     private Map<String, String> extraDataMap;
 
-    public TradeStatistics2(OfferPayload offerPayload,
+    public TradeStatistics2(FeeTxOfferPayload offerPayload,
                             Price tradePrice,
                             Coin tradeAmount,
                             Date tradeDate,
